hotfix: moved injured player information from playerinfo.py to home.py
